eth/tracers: remove duktape tracer and deps
