+++ eth/tracers/internal/tracetest/calltrace_test.go
@@ -134,10 +134,6 @@ func TestCallTracerNative(t *testing.T) {
 	testCallTracer("callTracer", "call_tracer", t)
 }
 
-func TestCallTracerLegacyDuktape(t *testing.T) {
-	testCallTracer("callTracerLegacyDuktape", "call_tracer_legacy", t)
-}
-
 func testCallTracer(tracerName string, dirPath string, t *testing.T) {
 	files, err := os.ReadDir(filepath.Join("testdata", dirPath))
 	if err != nil {
